moved functions to profiling/interview utils
