+++ user-study/gpt4-responses/v2_interview/v2_interview.py
@@ -1,18 +1,20 @@
 import openai
 import os
 import pandas as pd
-import random
-import os
 
 from v2_profiling import (
     UserProfile,
-    Auklets
+    Auklets,
+    create_userprofiles
 )
 
 from v2_interview_util import (
     get_msg,
     get_msg_with_image,
-    save_result_df
+    save_result_df,
+    select_questions,
+    read_questions,
+    read_human_data
 )
 
 from v2_interview_util_prompts import (
@@ -22,92 +24,6 @@ from v2_interview_util_prompts import (
 
 # openai.api_key = os.environ["OPENAI_API_KEY"]
 
-def read_human_data(path_to_csv:str, n=5, selection='first') -> pd.DataFrame:
-    """Reads data of max(n, length of csv file) users from given CSV file into DataFrame.
-        
-    Args:
-        path_to_csv (str) : path to a csv file with user data
-        n (int) : the number of users to create profiles for
-        selection (str) : how to select the n users from the dataset
-    
-    Returns:
-        ([UserProfile]) : a list of n UserProfile objects
-    """
-    selection_methods = ['random', 'first', 'last']
-
-    if selection not in selection_methods:
-        raise ValueError("Invalid selection method. Expected one of: %s" % selection_methods)
-    
-    if selection == 'first':
-        # if we only want the first n rows, we don't need to read the whole file
-        df = pd.read_csv(path_to_csv, nrows=n)
-        
-    elif selection == 'last':
-        df = pd.read_csv(path_to_csv)
-        df = df.tail(n)
-    else: 
-        df = pd.read_csv(path_to_csv)
-        df = df.sample(n)
-    
-    df['id'] = df.index
-    df = df.rename(columns={"Q_8" : "Q_4"})
-
-    return df
-    
-def create_userprofiles(data : pd.DataFrame) -> [UserProfile]:
-    """Constructs UserProfiles based on dataframe
-    
-    Args:
-        data (pd.DataFrame) : dataframe with human data
-        
-    Returns:
-        ([UserProfile]) : a list of UserProfile objects
-    """
-    return (data.apply(lambda x: UserProfile(x), axis = 1)).to_list()
-
-def read_questions(path_to_csv:str, indices:[int]) -> [(int, str)]:
-    """Reads file paths associated with questions from csv file
-    
-    Args:
-        path_to_csv (str) : path to csv file with file paths for each question
-        indices ([int]) : list of question indices (starting with 1)
-        
-    Returns:
-        ([(int, str)]) : list of index-filepath pairs
-    """
-    questions_df = pd.read_csv(path_to_csv)
-    return [(i, questions_df.at[i-1, "image_path"]) for i in indices]
-
-def select_questions(n=1, method='balanced') -> [int]:
-    """Selects n question indices in [1..20] using the specified selection method (default: 'balanced').
-    balanced: 1, 6, 11, 16, 2, 7, 12, 17, ...
-    random: as expected
-    first: 1, 2, 3, 4...
-
-    Args:
-        n (int) : number of questions to select
-        method (string) : how to select the questions. options: balanced, random, first
-    
-    Returns:
-        ([int]) : a list of question indices
-    """
-    
-    selection_methods = ['random', 'balanced', 'first']
-
-    if method not in selection_methods:
-        raise ValueError("Invalid selection method. Expected one of: %s" % selection_methods)
-    
-    n = max(n, 20)
-
-    # select number_questions out of 20 using selected method
-    if method == 'balanced':
-        question_IDs = [(i%4)*5 + (i//4) + 1 for i in range(n)]
-    elif method == 'random':
-        question_IDs = random.sample(range(1, 21), n)
-    else:
-        question_IDs = range(1, n+1) 
-
-    return question_IDs
 
 def single_interview(user : UserProfile, image_path : str, user_num : int, q_num : int) -> str:
     """
@@ -178,12 +94,15 @@ def simulate_interviews(number_users=1, number_questions=1, user_select='first',
             print(user_id)
             results_df.loc[user_id] = 'NA'
 
+        # profiling prompt only needs to be created once per UserProfile object
+        if user.profiling_prompt is None:
+            user.personalize_prompt(SYSTEM, profiling=True)
+
         # request gpt-4 responses for not yet (properly) answered questions
         for (index, q_path) in question_paths:
             question = "LLM_Q" + str(index) # TODO: will have to change this probably
             print(results_df.at[user_id, question])
             if results_df.at[user_id, question].lower() not in birds:
-                user.personalize_prompt(SYSTEM, profiling=True)
                 try:
                     results_df.at[user_id, question] = single_interview(user, q_path, user_num, index)
                 except:
